@@ -3,7 +3,7 @@ package com.example.androidstudio2dgamedevelopment.gamepanel;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Paint;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 
 import com.example.androidstudio2dgamedevelopment.R;
 
