@@ -7,7 +7,7 @@ import com.example.androidstudio2dgamedevelopment.GameDisplay;
 /**
  * GameObject is an abstract class which is the foundation of all world objects in the game.
  */
-public abstract class GameObject {
+public abstract class  GameObject {
     protected double positionX, positionY = 0.0;
     protected double velocityX, velocityY = 0.0;
     protected double directionX = 1.0;
